// src/lib/wiring/config/terminals.generated.ts
// AUTO-GENERATED from annotated hardware photos.
// Percentage coordinates relative to each cropped device image (x%, y%).
export const IMAGE_TERMINALS: Record<
  string,
  Record<string, [number, number]>
> = {
  PLC_FX3U: {
    GND: [28.53, 11.47],
    "S/S": [32.38, 11.47],
    "0V": [36.37, 11.42],
    X0: [40.22, 11.42],
    X2: [43.97, 11.42],
    X4: [47.72, 11.37],
    X6: [51.39, 11.37],
    X10: [55.08, 11.31],
    X12: [58.83, 11.37],
    X14: [62.63, 11.37],
    X16: [66.43, 11.42],
    X20: [70.23, 11.26],
    X22: [74.19, 11.26],
    X24: [78.07, 11.26],
    X26: [81.89, 11.21],
    L: [25.18, 15.66],
    N: [29.03, 15.5],
    "24V": [37.03, 15.17],
    X1: [40.99, 15.07],
    X3: [45.0, 15.01],
    X5: [48.96, 14.91],
    X7: [53.05, 14.42],
    X11: [57.01, 14.42],
    X13: [61.07, 14.42],
    X15: [65.08, 14.37],
    X17: [69.12, 14.32],
    X21: [73.13, 14.21],
    X23: [77.22, 14.16],
    X25: [81.31, 14.21],
    X27: [85.38, 14.16],
    Y0: [27.1, 86.38],
    Y2: [31.14, 86.38],
    Y4: [39.25, 86.49],
    Y6: [43.2, 86.49],
    Y10: [51.25, 86.76],
    Y12: [55.34, 86.7],
    Y14: [63.74, 86.43],
    Y16: [67.75, 86.27],
    Y20: [71.81, 86.54],
    Y22: [75.9, 86.54],
    Y24: [80.05, 86.6],
    Y26: [84.22, 86.49],
    COM5: [88.52, 86.43],
    COM1: [26.15, 91.21],
    Y1: [30.17, 91.15],
    Y3: [34.13, 91.05],
    COM2: [38.03, 91.26],
    Y5: [41.81, 91.1],
    Y7: [45.76, 91.21],
    COM3: [49.67, 91.26],
    Y11: [53.55, 91.21],
    Y13: [57.4, 91.15],
    COM4: [61.23, 91.1],
    Y15: [65.19, 91.1],
    Y17: [69.17, 91.21],
    Y21: [73.11, 91.1],
    Y23: [77.09, 91.05],
    Y25: [81.02, 90.99],
    Y27: [85.01, 91.05],
  },
  PLC_FX5U: {
    // Top strip, upper row (X even + GND/S-S/0V)
    GND: [34, 7.98],
    "S/S": [38.5, 7.99],
    "0V": [43.5, 8.02],
    X0: [48.5, 7.89],
    X2: [53.5, 7.87],
    X4: [58, 7.87],
    X6: [63, 7.9],
    X10: [67.5, 7.98],
    X12: [72.5, 7.9],
    X14: [77, 7.86],
    X16: [81.5, 7.86],
    // Top strip, lower row (X odd + L/N/24V)
    L: [31.5, 13.54],
    N: [36.5, 13.43],
    "24V": [46, 13.45],
    X1: [51, 13.48],
    X3: [56, 13.38],
    X5: [61, 13.54],
    X7: [66, 13.51],
    X11: [70.5, 13.55],
    X13: [75.5, 13.49],
    X15: [80.5, 13.49],
    X17: [85.5, 13.57],

    // Bottom strip, upper row (Y even)
    Y0: [33.5, 85],
    Y2: [38.5, 85],
    Y4: [48, 85],
    Y6: [53, 85],
    Y10: [63, 85],
    Y12: [68, 85],
    Y14: [77.5, 85],
    Y16: [82.5, 85],
    // Bottom strip, lower row (Y odd + COM0-3)
    COM0: [31, 92],
    Y1: [36, 92],
    Y3: [41, 92],
    COM1: [45.5, 92],
    Y5: [50.5, 92],
    Y7: [55.05, 92],
    COM2: [60, 92],
    Y11: [64.5, 92],
    Y13: [69.5, 92],
    COM3: [74, 92],
    Y15: [79, 92],
    Y17: [84, 92],

    // Left-side analog I/O block (tight ~2.2% pitch)
    "V1+AD": [11.3, 27],
    "V2+AD": [13.7, 27],
    "V-AD": [16.2, 27],
    "V+DA": [18.4, 27],
    "V-DA": [20.5, 27],

    // Left-side RS-485 block
    SG: [11, 75.5],
    SDB: [13.5, 75.5],
    SDA: [16, 75.5],
    RDB: [18.5, 75.5],
    RDA: [20.65, 75.5],
  },
  MCB: {
    L1_IN: [25.22, 15.66],
    L2_IN: [74.96, 15.8],
    L1_OUT: [24.52, 86.17],
    L2_OUT: [73.2, 86.31],
  },
  PSU: {
    L_IN: [24.52, 8.64],
    N_IN: [54.09, 8.64],
    "0VDC": [22.6, 90.79],
    "24VDC": [77.4, 90.88],
  },
};